+++ examples/order-python-flask/app/core/config.py
@@ -1,6 +1,0 @@
-class Config:
-    TESTING = False
-
-
-class TestingConfig(Config):
-    TESTING = True